+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#233](https://github.com/os2display/display-admin-client/pull/233)
+  - Cleaned up code flow in playlist saving.
+  - Changed to chaining relations puts.
 - [#232](https://github.com/os2display/display-admin-client/pull/232)
   - Fixed time zone issue in playlist schedules.
 - [#231](https://github.com/os2display/display-admin-client/pull/231)